+++ src/API/Tests/Lifo/Pop.c
@@ -32,7 +32,7 @@ int main( int argc, char * argv[] )
    const char * data1 = "My Data1";
    const char * data2 = "My Data 2";
    char buffer[200];
-   size_t length;
+   uint32_t length;
    psoObjectDefinition defLilo = { 
       PSO_LIFO,
       1,
