+++ src/API/Tests/Map/GetFirst.c
@@ -32,7 +32,7 @@ int main( int argc, char * argv[] )
    const char * data = "My Data";
    char buffer[200];
    char buffKey[50];
-   size_t dataLength, keyLength;
+   uint32_t dataLength, keyLength;
    psoObjectDefinition mapDef = { 
       PSO_FAST_MAP, 
       1,
