+++ src/C++/Tests/HashMap/GetNext.cpp
@@ -38,7 +38,7 @@ int main( int argc, char * argv[] )
    const char * data2 = "My Data2";
    char buffer[50];
    char buffKey[50];
-   size_t dataLength, keyLength;
+   uint32_t dataLength, keyLength;
    psoObjectDefinition folderDef;
    psoObjectDefinition mapDef = { 
       PSO_HASH_MAP,
